+++ src/app/core/models/user.model.ts
@@ -22,15 +22,3 @@ export interface LoginPayload {
   password: string;
 }
 
-export interface ForgotPasswordPayload {
-  email: string;
-}
-
-export interface ResetPasswordPayload {
-  token: string;
-  password: string;
-}
-
-export interface MessageResponse {
-  message: string;
-}